Rewrite the spell suggestion plumbing and split out the z= prompt
`spellsuggest.rs` was the last transpiled file of the suggestion
subsystem: 1,272 lines over the 1,000-line cap, and the only file of its
family without `#![deny(unsafe_op_in_unsafe_fn)]`. It now holds only the
plumbing — the `'spellsuggest'` option, `spell_find_suggest` and the
methods it dispatches to — and the interactive `z=` command moved to
`spellsuggest/prompt.rs`.

## The split

`spell_suggest` is the one piece with no bearing on the rest: it finds
the bad word, draws a numbered list, asks, and edits the line. Nothing
else in the tree needs any of it, so `prompt.rs` takes it along with the
constants only it used (`ESC`, `IOSIZE`).

Its C had a `goto skip` restoring `curwin->w_p_spell` on every exit path.
That is a scope guard, and it is now written as one: `spell_suggest` does
nothing but save the option, force it on, call `suggest_and_replace`, and
put it back — so the body is free to return early where the C jumped, and
the three early exits (no `'spelllang'`, multi-line Visual selection, no
word at the cursor) cannot skip the restore.

## Constants got real types

The transpiled preamble carried a dead `pub use` blob and a dozen
`C2Rust_Unnamed_*` aliases, all `c_uint`, whose every consumer wanted
`c_int`. Typing the `SCORE_*`, `WF_*` and `SPS_*` constants `c_int`, and
`MAXWLEN` `usize` (23 of its 36 uses were array sizes), retired ~150
casts across the family. `WC_KEY_OFF` and `sftword_T` moved to the single
module that uses each, and both are now `offset_of!` rather than a
hand-written 2. `limitscore_T` was dead and is gone, as are `NULL`,
`true_0`, `false_0` and `DEFAULT_MAXPATHL`.

The submodules are now private `mod`s: only `spell_suggest`,
`spell_suggest_list` and `spell_check_sps` are used outside the family,
and only those are `pub` from the parent now.

## Things that had to stay exactly as they were

- `spell_find_suggest` picks the default sound-folding language out of
  `curbuf`'s language list, not `curwin`'s, so it cannot use
  `window_langs()` like its neighbours. Called out in a comment.
- The three sound-a-like passes read as a loop over rising ceilings, but
  the first one runs unconditionally where the later two are guarded on
  the list being short. Folding all three into one loop changed
  behaviour, because the list can already be at the clean-up count when
  the first pass starts.
- Pointers into `su`'s inline word buffers are taken with `&raw`, not
  `as_mut_ptr()`. `suggest_try_special` plants a terminator in
  `su_fbadword` *between* deriving the pointer and using it, which would
  invalidate a `&mut`-derived one.

## Verification

Differential sweep over `:mkspell`, `:spellinfo`, `spellbadword()` and
`spellsuggest()` (30 corpus cases, 35 generated `.spl`/`.sug`): report
and file hashes identical to the pre-rewrite binary. The sweep does not
drive `z=`, so a second probe scripted 23 prompt scenarios — counted and
interactive `z=`, choice 0/1/2, `'verbose'` scores, `'rightleft'`
mirroring, Visual selection forwards and backwards, multi-line Visual,
no word at the cursor, `'spell'` off, empty `'spelllang'`,
`:spellrepall`, redo with `.`, and `'spellsuggest'` best/fast/double/limit
— and diffed the listings, buffers, cursor columns and messages against
the old binary byte for byte.

`just oldtest test_spell test_spell_utf8 test_spellfile test_spellrare`,
the four spell functional specs (26 passed) and the same specs under
AddressSanitizer all pass with no reports.

`just refresh --allow-growth`: `prompt.rs` is a new file, so its 7
`unsafe` blocks and its lack of `#![forbid(unsafe_code)]` read as growth.
The parent lost its over-cap line count and two `unsafe` blocks, and its
33 clippy warnings went to zero.

+++ crates/nvim/src/nvim/spellsuggest/walk.rs
@@ -91,11 +91,11 @@ use crate::src::nvim::types::{idx_T, int64_t, langp_T, proftime_T, slang_T};
 use core::ffi::{c_char, c_int};
 
 /// One level per byte of the bad word is all the walk can ever need.
-const STACK_SIZE: usize = MAXWLEN as usize;
+const STACK_SIZE: usize = MAXWLEN;
 
 /// `preword` holds a prefix, any number of compounded words and a split
 /// word one after another, so it needs room for several whole words.
-const PREWORD_SIZE: usize = MAXWLEN as usize * 3;
+const PREWORD_SIZE: usize = MAXWLEN * 3;
 
 // Values for `Frame::prefix_depth` that are not a stack depth.
 /// This level is not below a postponed prefix.
@@ -273,13 +273,13 @@ pub(crate) struct Walk {
     pub repextra: c_int,
 
     /// The good word collected from the tree so far.
-    pub tword: [c_char; MAXWLEN as usize],
+    pub tword: [c_char; MAXWLEN],
     /// The suggestion with its proper case: prefix, compounded words and
     /// split words concatenated. NUL-terminated while going deeper, but
     /// not on the way back up.
     pub preword: [c_char; PREWORD_SIZE],
     /// One compound flag per word collected in `preword`.
-    pub compflags: [u8; MAXWLEN as usize],
+    pub compflags: [u8; MAXWLEN],
 
     pub stack: [Frame; STACK_SIZE],
     pub depth: c_int,
@@ -346,9 +346,9 @@ impl Walk {
                 idxs: core::ptr::null_mut(),
                 fword,
                 repextra: 0,
-                tword: [0; MAXWLEN as usize],
+                tword: [0; MAXWLEN],
                 preword: [0; PREWORD_SIZE],
-                compflags: [0; MAXWLEN as usize],
+                compflags: [0; MAXWLEN],
                 stack: [Frame::default(); STACK_SIZE],
                 depth: 0,
                 breakcheckcount: 1000,
